+++ cmd/reloader/template/main.go
@@ -17,7 +17,7 @@ import (
 	cfgcore "github.com/apecloud/kubeblocks/internal/configuration"
 	cfgcm "github.com/apecloud/kubeblocks/internal/configuration/config_manager"
 	cfgutil "github.com/apecloud/kubeblocks/internal/configuration/util"
-	"github.com/apecloud/kubeblocks/internal/controller/configuration"
+	"github.com/apecloud/kubeblocks/internal/controller/configfsm"
 	"github.com/apecloud/kubeblocks/internal/gotemplate"
 )
 
@@ -104,7 +104,7 @@ func main() {
 		failed(err, "failed to parse config spec")
 	}
 
-	mergePolicy, err := configuration.NewTemplateMerger(*configRenderMeta.LegacyRenderedConfigSpec,
+	mergePolicy, err := configfsm.NewTemplateMerger(*configRenderMeta.LegacyRenderedConfigSpec,
 		context.TODO(), nil, nil, *configRenderMeta.ComponentConfigSpec, &appsv1alpha1.ConfigConstraintSpec{
 			FormatterConfig: &configRenderMeta.FormatterConfig,
 		})
