+++ internal/controller/configfsm/builtin_env_test.go
@@ -17,7 +17,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package configuration
+package configfsm
 
 import (
 	"strconv"
