+++ internal/controller/configfsm/template_merger.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package configuration
+package configfsm
 
 import (
 	"context"
