+++ internal/controller/configfsm/tool_image_builder_test.go
@@ -17,7 +17,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package configuration
+package configfsm
 
 import (
 	. "github.com/onsi/ginkgo/v2"
